swagger_server/controllers/upload_sensor_data_controller.py
import connexion
import six
from flask import jsonify
from swagger_server import util
import pandas as pd
import json

def upload_post(upfile):  # noqa: E501

    df = pd.read_csv(upfile)
    df.to_csv("/tmp/iot.csv")
    data = {"message": "File Successfully uploaded"}

    return data, 200