requirements.txt
connexion[swagger-ui]
python_dateutil == 2.6.0
setuptools >= 21.0.0
flask_cors